add a screensaver
Currently set to ~1min until activation, see screen_saver.c - ACTIVE_AFTER.

The screensaver is reset with any touch or any screen_stack
change. The latter means there is something new for the user to see.

API calls per se do not reset the screensaver - only if they display
something via screen_stack.

The screensaver currently also appears if the user is in the middle of
a workflow or API call, if they do not interact with the device. It is
unclear if this is desirable, but in general it would work well
together with a general API timeout or automatic device lock.

Animation: logo scrolls from left ouf of screen to right out of
screen, and wraps around one pixel lower. For a short while before the
logo is visible, the screen is black, so that the change to the
screensaver is not too quick.

Finally, some more thinking is required around the three main screens
and when and how to present them:

 1. "See the BitBoxApp" - currently shown until unlocked.
 2. Default waiting screen when nothing is going on. Shows the
      BitBox02 logo.
 3. Screensaver after inactivity. Currently it is an animated
      BitBox02 logo.

Currently it is confusing to leave the animated screensaver just to go
back to the same image logo.