workflow/restore_from_mnemonic: restrict choices for the last word
When one has entered 23 words, the valid choices for the
last word can be restricted to a subset of all bip39 words for which
the seedphrase checksum is valid.

This makes the entry of the last word a lot quicker 8 choices instead
of 2048.

For 12/18 words, there are a lot more valid words, so we cannot show
them in a menu. We could restrict keyboard entry to these choices, but
it has a big UX downside: if there is a typo in the words before,
you'd get stuck in the last word, as your word is not present, and
there would be no message about invalid recovery words. This is why we
only enable this feature for 24 words. In this case, all choices can
be shown in a menu with an additional entry 'None of them'.

This also enables an expert user to manually create and enter a seed
using e.g. dice throws, without using other software to compute the
checksum. They can pick an arbitrary word instead. This leads to an
entropy reduction from 256 bits to 253 bits for 24 words, which is not
a problem.