rust/api: port RestoreBackup api call to Rust
It is very similar to RestoreFromMnemonic. The next commit will unify
the two functions.

Note that the return error values change from Generic to UserAbort,
Memory, etc. depending on the actual error. Same as in
4d88865ae85021fafc82420d96548fa0d771fd95 where the same happened to
RestoreFromMnemonic, this is technically a breaking change, but we
will not bump the firmware major version for this. The BitBoxApp will
handle either error the same way. No known downstream user uses the
Python function except for send_message.py, where we will fix it up in
another commit.