qtouch: subtract min sensor value per slider
This improves touch position accuracy in noisy environments
and may reduce touch issues reported in the wild.

Touch position is calculated with a weighted average of the sensor
readings. If properly calibrated, sensors on the opposite end of a
finger touch would be zero and thus make no contribution to the
weighted average. If the baseline sensor readings are elevated, the
sensors on the opposite edge DO contribute to the weighted average
making a positional artifact (i.e. the position is more central than
it should be in reality). This artifact is higher when the finger is
a bit distant while approaching and lower/negligible when the finger
is fully touching the device. This can cause the position to move
enough to enter "slide" mode and disable "tap" events being emitted.