ui/trinary_input_string: specialize to bip39 wordlist
The C implementation accepted any list of strings to use for the
wordlist in the keyboard. So far we've only used the English BIP39
wordlist to restrict user input to valid words when restoring from
mnemonic.

The Rust wrapper already specialized on BIP39 only. It loaded all
words at once (2048 mallocs) incl. null terminator so it can be used
by C.

In this commit we change the C code to directly get the bip39 word by
index, skipping all the boilerplate in Rust.

Apart from using much less heap memory, this has one big benefit:

We want to restrict the 12th / 18th word to the valid checksum
candidate words, just like with the 24th word, but since they have
many more candidates, we need to use the regular keyboard input (the
24th word choice is a simple menu with 8 entries). This was not
previously possible because the Rust wrapper forced using all BIP39
words. Now it is easy as one can simply provide the indices of the
candidate words.