keystore: more consistent treatment of keystore_unlock errors
Before, the device would be reset right after the 10th wrong
password. However, if there was any other error during unlock (e.g. a
securechip error) that is not caused by a wrong password, the device
would not be reset immediately, but at the start of the next unlock
attempt.

This commit unifies all error treatment, leading to a device reset
right away.

The alternative is to decrement the counter on any error other than
incorrect password, but that would drastically change the behavior of
the unlock function, and we'd rather be conservative here.